unify scalar constructors
