+++ vortex-scalar/src/scalar.rs
@@ -10,20 +10,12 @@ use std::hash::Hasher;
 use vortex_dtype::DType;
 use vortex_dtype::NativeDType;
 use vortex_dtype::PType;
-use vortex_error::VortexExpect;
 use vortex_error::VortexResult;
 use vortex_error::vortex_ensure;
+use vortex_error::vortex_ensure_eq;
 
-use crate::BinaryScalar;
-use crate::BoolScalar;
-use crate::DecimalScalar;
-use crate::ExtScalar;
-use crate::ListScalar;
 use crate::PValue;
-use crate::PrimitiveScalar;
 use crate::ScalarValue;
-use crate::StructScalar;
-use crate::Utf8Scalar;
 
 /// A typed scalar value.
 ///
@@ -90,11 +82,15 @@ impl Scalar {
 
     /// Creates a new [`Scalar`] with the given [`DType`] and potentially null [`ScalarValue`].
     ///
+    /// This is just a helper function for tests.
+    ///
     /// # Panics
     ///
     /// Panics if the given [`DType`] and [`ScalarValue`] are incompatible.
     #[cfg(test)]
     pub fn new(dtype: DType, value: Option<ScalarValue>) -> Self {
+        use vortex_error::VortexExpect;
+
         Self::try_new(dtype, value).vortex_expect("Failed to create Scalar")
     }
 
@@ -273,6 +269,32 @@ impl Scalar {
         self.value.is_none()
     }
 
+    /// Reinterprets the bytes of this scalar as a different primitive type.
+    ///
+    /// # Errors
+    ///
+    /// Panics if the scalar is not a primitive type or if the types have different byte widths.
+    pub fn primitive_reinterpret_cast(&self, ptype: PType) -> VortexResult<Self> {
+        let primitive = self.as_primitive();
+        if primitive.ptype() == ptype {
+            return Ok(self.clone());
+        }
+
+        vortex_ensure_eq!(
+            primitive.ptype().byte_width(),
+            ptype.byte_width(),
+            "can't reinterpret cast between integers of two different widths"
+        );
+
+        Scalar::try_new(
+            DType::Primitive(ptype, self.dtype().nullability()),
+            primitive
+                .pvalue()
+                .map(|p| p.reinterpret_cast(ptype))
+                .map(ScalarValue::Primitive),
+        )
+    }
+
     /// Returns the size of the scalar in bytes, uncompressed.
     #[cfg(test)]
     pub fn nbytes(&self) -> usize {
@@ -308,135 +330,6 @@ impl Scalar {
                 .unwrap_or_default(),
             DType::Extension(_) => self.as_extension().storage().nbytes(),
         }
-    }
-}
-
-/// Scalar downcasting methods to typed views.
-impl Scalar {
-    /// Returns a view of the scalar as a boolean scalar.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the scalar does not have a [`DType::Bool`] type.
-    pub fn as_bool(&self) -> BoolScalar<'_> {
-        self.as_bool_opt()
-            .vortex_expect("Failed to convert scalar to bool")
-    }
-
-    /// Returns a view of the scalar as a boolean scalar if it has a boolean type.
-    pub fn as_bool_opt(&self) -> Option<BoolScalar<'_>> {
-        BoolScalar::try_new(self.dtype(), self.value()).ok()
-    }
-
-    /// Returns a view of the scalar as a primitive scalar.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the scalar does not have a [`DType::Primitive`] type.
-    pub fn as_primitive(&self) -> PrimitiveScalar<'_> {
-        self.as_primitive_opt()
-            .vortex_expect("Failed to convert scalar to primitive")
-    }
-
-    /// Returns a view of the scalar as a primitive scalar if it has a primitive type.
-    pub fn as_primitive_opt(&self) -> Option<PrimitiveScalar<'_>> {
-        PrimitiveScalar::try_new(self.dtype(), self.value()).ok()
-    }
-
-    /// Returns a view of the scalar as a decimal scalar.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the scalar does not have a [`DType::Decimal`] type.
-    pub fn as_decimal(&self) -> DecimalScalar<'_> {
-        self.as_decimal_opt()
-            .vortex_expect("Failed to convert scalar to decimal")
-    }
-
-    /// Returns a view of the scalar as a decimal scalar if it has a decimal type.
-    pub fn as_decimal_opt(&self) -> Option<DecimalScalar<'_>> {
-        DecimalScalar::try_new(self.dtype(), self.value()).ok()
-    }
-
-    /// Returns a view of the scalar as a UTF-8 string scalar.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the scalar does not have a [`DType::Utf8`] type.
-    pub fn as_utf8(&self) -> Utf8Scalar<'_> {
-        self.as_utf8_opt()
-            .vortex_expect("Failed to convert scalar to utf8")
-    }
-
-    /// Returns a view of the scalar as a UTF-8 string scalar if it has a UTF-8 type.
-    pub fn as_utf8_opt(&self) -> Option<Utf8Scalar<'_>> {
-        Utf8Scalar::try_new(self.dtype(), self.value()).ok()
-    }
-
-    /// Returns a view of the scalar as a binary scalar.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the scalar does not have a [`DType::Binary`] type.
-    pub fn as_binary(&self) -> BinaryScalar<'_> {
-        self.as_binary_opt()
-            .vortex_expect("Failed to convert scalar to binary")
-    }
-
-    /// Returns a view of the scalar as a binary scalar if it has a binary type.
-    pub fn as_binary_opt(&self) -> Option<BinaryScalar<'_>> {
-        BinaryScalar::try_new(self.dtype(), self.value()).ok()
-    }
-
-    /// Returns a view of the scalar as a struct scalar.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the scalar does not have a [`DType::Struct`] type.
-    pub fn as_struct(&self) -> StructScalar<'_> {
-        self.as_struct_opt()
-            .vortex_expect("Failed to convert scalar to struct")
-    }
-
-    /// Returns a view of the scalar as a struct scalar if it has a struct type.
-    pub fn as_struct_opt(&self) -> Option<StructScalar<'_>> {
-        StructScalar::try_new(self.dtype(), self.value()).ok()
-    }
-
-    /// Returns a view of the scalar as a list scalar.
-    ///
-    /// Note that we use [`ListScalar`] to represent **both** [`DType::List`] and
-    /// [`DType::FixedSizeList`].
-    ///
-    /// # Panics
-    ///
-    /// Panics if the scalar does not have a [`DType::List`] or [`DType::FixedSizeList`] type.
-    pub fn as_list(&self) -> ListScalar<'_> {
-        self.as_list_opt()
-            .vortex_expect("Failed to convert scalar to list")
-    }
-
-    /// Returns a view of the scalar as a list scalar if it has a list type.
-    ///
-    /// Note that we use [`ListScalar`] to represent **both** [`DType::List`] and
-    /// [`DType::FixedSizeList`].
-    pub fn as_list_opt(&self) -> Option<ListScalar<'_>> {
-        ListScalar::try_new(self.dtype(), self.value()).ok()
-    }
-
-    /// Returns a view of the scalar as an extension scalar.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the scalar does not have a [`DType::Extension`] type.
-    pub fn as_extension(&self) -> ExtScalar<'_> {
-        self.as_extension_opt()
-            .vortex_expect("Failed to convert scalar to extension")
-    }
-
-    /// Returns a view of the scalar as an extension scalar if it has an extension type.
-    pub fn as_extension_opt(&self) -> Option<ExtScalar<'_>> {
-        ExtScalar::try_new(self.dtype(), self.value()).ok()
     }
 }
